Refactor monolithic app.py into modular structure (core/ui/config) and clean .gitignore
