+++ .gitignore
@@ -1,1 +1,58 @@
-<paste that content>
+# --- Python / FastAPI standard ignores ---
+__pycache__/
+*.py[cod]
+*.pyo
+*.pyd
+*.so
+*.egg
+*.egg-info/
+.eggs/
+dist/
+build/
+develop-eggs/
+.installed.cfg
+*.manifest
+*.spec
+
+# --- Virtual environments ---
+.venv/
+venv/
+ENV/
+env/
+pip-wheel-metadata/
+
+# --- Streamlit ---
+.streamlit/
+
+# --- IDE / OS files ---
+.DS_Store
+.idea/
+.vscode/
+*.swp
+.AppleDouble
+.LSOverride
+Icon?
+._*
+
+# --- Environment variables ---
+.env
+.env.*
+!.env.example
+
+# --- Logs / temporary ---
+*.log
+logs/
+tmp/
+
+# --- Node / Frontend (if you use Vite/React) ---
+node_modules/
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+package-lock.json
+pnpm-lock.yaml
+
+# --- Jupyter ---
+.ipynb_checkpoints
+
+
