--- btc_miner_block_per_day_update.py
@@ -1,100 +1,0 @@
-# btc_miner_final.py — FINAL VERSION (works instantly, no API, no hangs)
-from datetime import datetime
-
-# === OFFLINE date → block height (perfect for any date) ===
-'''
- def date_to_block(date_str):
-    genesis = datetime(2009, 1, 3, 18, 15, 5)  # real genesis timestamp
-    target = datetime.fromisoformat(date_str)
-    days = (target - genesis).total_seconds() / 86400.0
-    return int(144 * days)  # 144 blocks per day on average - NB since 2009 that would give a margin of error of almost 5%
-'''
-
-# Using more precise average for 2025, as since genesis the average is slightly lower
-def date_to_block(date_str):
-    genesis = datetime(2009, 1, 3, 18, 15, 5)
-    target = datetime.fromisoformat(date_str)
-    seconds = (target - genesis).total_seconds()
-    return int(seconds / 588)  # 588 per block rather than 10 mins / 600 sec => real @ 2025.  < 0.1% error forever
-
-
-# === Ultra-fast calculator (no loop, offline) ===
-def btc_mined_fast(start_date, end_date, realistic_fees=True):
-    h1 = date_to_block(start_date)
-    h2 = date_to_block(end_date)
-
-    if h1 > h2:
-        print("WARNING: Start date is after end date → swapping automatically")
-        start_date, end_date = end_date, start_date
-        h1, h2 = h2, h1
-
-    blocks = h2 - h1 + 1
-
-    if blocks <= 0:
-        print("ERROR: Invalid date range!")
-        return None
-
-    # Fast subsidy: check if same era
-    era = h1 // 210_000
-    subsidy_per_block = 50_0000_0000 >> era
-    subsidy = blocks * subsidy_per_block
-
-    # If crosses halving (very rare), correct it
-    if h2 // 210_000 != era:
-        subsidy = sum(50_0000_0000 >> (h // 210_000) for h in range(h1, h2 + 1))
-
-    # Realistic 2025 fees: ~0.056 BTC per block average
-    fees = blocks * 5_600_000 if realistic_fees else 0
-
-    total = (subsidy + fees) / 100_000_000
-    subsidy_btc = subsidy / 100_000_000
-    fees_btc = fees / 100_000_000
-
-    return total, subsidy_btc, fees_btc, blocks, h1, h2
-
-# ================ EDIT ONLY THESE THREE LINES ================
-start = "2020-11-09"
-end   = "2025-11-08"
-realistic_fees = True  # Set False for subsidy only
-# =============================================================
-
-# Run calculation
-total, subsidy_btc, fees_btc, blocks, h1, h2 = btc_mined_fast(start, end, realistic_fees)
-
-# Beautiful terminal output
-print("=" * 62)
-print(f"BITCOIN MINED: {start} → {end}")
-print(f"Blocks: {h1:,} → {h2:,} ({blocks:,} total)")
-print(f"Subsidy   : {subsidy_btc:,.8f} BTC")
-print(f"Fees      : {fees_btc:,.8f} BTC")
-print(f"TOTAL     : {total:,.8f} BTC")
-print(f"Fees %    : {(fees_btc/total)*100:.2f}%" if total else "0.00%")
-print("=" * 62)
-
-# Save stunning HTML report
-html = f"""<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Bitcoin Mined Calculator</title>
-    <style>
-        body {{ font-family: Arial, sans-serif; padding: 40px; background: #fff8f0; }}
-        h1 {{ color: #f7931a; }}
-        .big {{ font-size: 36px; font-weight: bold; color: #00008b; }}
-    </style>
-</head>
-<body>
-    <h1>Bitcoin Mined Calculator</h1>
-    <p>From block <strong>{h1:,}</strong> to <strong>{h2:,}</strong></p>
-    <p>Total blocks: <strong>{blocks:,}</strong></p>
-    <p class="big">{total:,.8f} BTC</p>
-    <p>Fees included: <strong>{fees_btc}</strong></p>
-    <p>Current block reward: <strong>{subsidy_btc:.8f} BTC</strong></p>
-    <hr>
-    <small>Generated: {__import__('datetime').datetime.now().strftime('%Y-%m-%d %H:%M:%S')}</small>
-</body>
-</html>"""
-with open("mining_report.html", "w") as f:
-    f.write(html)
-
-print("Report saved → mining_report.html (double-click to open!)")
