--- btc_miner_blocks_and_fees_update.py
@@ -1,148 +1,0 @@
-# btc_miner_final.py — FINAL VERSION: 100% WORKING, CONFIGURABLE
-from datetime import datetime
-import requests
-
-# ========================================
-# CONFIGURABLE CONSTANTS — EDIT THESE ONLY
-# ========================================
-GENESIS_TIMESTAMP = datetime(2009, 1, 3, 18, 15, 5)
-BLOCK_TIME_SECONDS = 588  # Long-term observed average (more accurate than 600)
-SATOSHIS_PER_BTC = 100_000_000
-BASE_SUBSIDY_SATS = 50 * SATOSHIS_PER_BTC  # 50 BTC in satoshis
-HALVING_INTERVAL = 210_000
-FALLBACK_AVG_FEE_BTC = 0.061  # Safe fallback if API fails
-FEE_API_URL = "https://mempool.space/api/v1/blocks"
-FEE_BLOCKS_TO_AVERAGE = 2016  # ~2 weeks for stable average
-# ========================================
-
-
-def get_real_avg_fee_per_block() -> float:
-    """Fetch real average fee per block from mempool.space (last N blocks)."""
-    try:
-        response = requests.get(FEE_API_URL, timeout=15)
-        response.raise_for_status()
-        blocks = response.json()[:FEE_BLOCKS_TO_AVERAGE]
-        if not blocks:
-            raise ValueError("No blocks returned from API")
-
-        total_fees_sats = sum(block["extras"]["totalFees"] for block in blocks)
-        avg_fee_btc = total_fees_sats / len(blocks) / SATOSHIS_PER_BTC
-        print(f"Real avg fee fetched ({len(blocks)} blocks): {avg_fee_btc:.6f} BTC/block")
-        return avg_fee_btc
-    except Exception as e:
-        print(f"API failed ({e}) → using fallback {FALLBACK_AVG_FEE_BTC} BTC")
-        return FALLBACK_AVG_FEE_BTC
-
-
-def date_to_block(date_str: str) -> int:
-    """Convert ISO date (YYYY-MM-DD) to approximate block height."""
-    try:
-        target = datetime.fromisoformat(date_str + "T00:00:00")
-    except ValueError:
-        raise ValueError("Date must be in YYYY-MM-DD format")
-
-    seconds_since_genesis = (target - GENESIS_TIMESTAMP).total_seconds()
-    return int(seconds_since_genesis / BLOCK_TIME_SECONDS)
-
-
-def btc_mined_fast(start_date: str, end_date: str, realistic_fees: bool = True):
-    """Calculate total BTC mined (subsidy + fees) between two dates."""
-    h1 = date_to_block(start_date)
-    h2 = date_to_block(end_date)
-
-    if h1 > h2:
-        print("Swapping dates automatically")
-        start_date, end_date = end_date, start_date
-        h1, h2 = h2, h1
-
-    blocks = h2 - h1 + 1
-    if blocks <= 0:
-        print("ERROR: Invalid block range!")
-        return None
-
-    # Calculate subsidy per era
-    era1 = h1 // HALVING_INTERVAL
-    era2 = h2 // HALVING_INTERVAL
-
-    if era1 == era2:
-        subsidy_sats = blocks * (BASE_SUBSIDY_SATS >> era1)
-    else:
-        subsidy_sats = sum(
-            BASE_SUBSIDY_SATS >> (h // HALVING_INTERVAL)
-            for h in range(h1, h2 + 1)
-        )
-
-    # Fees
-    avg_fee_btc = get_real_avg_fee_per_block() if realistic_fees else 0.0
-    fees_sats = int(blocks * avg_fee_btc * SATOSHIS_PER_BTC)
-
-    # Totals
-    total_btc = (subsidy_sats + fees_sats) / SATOSHIS_PER_BTC
-    subsidy_btc = subsidy_sats / SATOSHIS_PER_BTC
-    fees_btc = fees_sats / SATOSHIS_PER_BTC
-
-    return (
-        total_btc,
-        subsidy_btc,
-        fees_btc,
-        blocks,
-        h1,
-        h2,
-        avg_fee_btc,
-        start_date,
-        end_date,
-    )
-
-
-# ================ EDIT THESE ================
-start = "2020-11-09"
-end = "2025-11-08"
-realistic_fees = True
-# ===========================================
-
-result = btc_mined_fast(start, end, realistic_fees)
-if result is None:
-    print("Failed.")
-else:
-    total, subsidy, fees, blocks, h1, h2, avg_fee, s_date, e_date = result
-    print("=" * 70)
-    print(f"BITCOIN MINED: {s_date} to {e_date}")
-    print(f"Blocks: {h1:,} to {h2:,} ({blocks:,} total)")
-    print(f"Subsidy : {subsidy:,.8f} BTC")
-    print(f"Fees (30d avg {avg_fee:.6f}) : {fees:,.8f} BTC")
-    print(f"TOTAL : {total:,.8f} BTC")
-    print(f"Fees % : {(fees / total) * 100:.2f}%" if total > 0 else "Fees % : N/A")
-    print("=" * 70)
-
-    # Generate HTML Report
-    html = f"""<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Mined {s_date} to {e_date}</title>
-    <style>
-        body {{font-family:system-ui;background:#fff8f0;padding:60px}}
-        .box {{background:white;padding:40px;border-radius:24px;box-shadow:0 20px 40px rgba(0,0,0,0.1);max-width:680px;margin:auto}}
-        h1 {{color:#f7931a;text-align:center}}
-        .big {{font-size:68px;font-weight:900;color:#00008b;text-align:center}}
-        .tag {{background:#f7931a;color:white;padding:6px 14px;border-radius:12px;font-size:0.9em}}
-    </style>
-</head>
-<body>
-<div class="box">
-<h1>Bitcoin Mining Revenue Report</h1>
-<p><strong>Period:</strong> {s_date} to {e_date}</p>
-<p><strong>Blocks:</strong> {h1:,} to {h2:,} ({blocks:,} total)</p>
-<div class="big">{total:,.8f} BTC</div>
-<p>Subsidy: {subsidy:,.8f} BTC<br>
-   Fees: <strong>{fees:,.8f} BTC</strong>
-   <span class="tag">Real {FEE_BLOCKS_TO_AVERAGE}-block avg: {avg_fee:.6f} BTC/block</span></p>
-<hr>
-<small>Generated • Fees from mempool.space • {datetime.now():%Y-%m-%d %H:%M:%S}</small>
-</div>
-</body>
-</html>"""
-
-    with open("mining_report.html", "w", encoding="utf-8") as f:
-        f.write(html)
-    print("Report saved → mining_report.html")
