--- mining_report.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="UTF-8">
-    <title>Mined 2020-11-09 to 2025-11-08</title>
-    <style>
-        body {font-family:system-ui;background:#fff8f0;padding:60px}
-        .box {background:white;padding:40px;border-radius:24px;box-shadow:0 20px 40px rgba(0,0,0,0.1);max-width:680px;margin:auto}
-        h1 {color:#f7931a;text-align:center}
-        .big {font-size:68px;font-weight:900;color:#00008b;text-align:center}
-        .tag {background:#f7931a;color:white;padding:6px 14px;border-radius:12px;font-size:0.9em}
-    </style>
-</head>
-<body>
-<div class="box">
-<h1>Bitcoin Mining Revenue Report</h1>
-<p><strong>Period:</strong> 2020-11-09 to 2025-11-08</p>
-<p><strong>Blocks:</strong> 635,839 to 904,002 (268,164 total)</p>
-<div class="big">1,481,334.25444256 BTC</div>
-<p>Subsidy: 1,476,015.62500000 BTC<br>
-   Fees: <strong>5,318.62944256 BTC</strong>
-   <span class="tag">Real 2016-block avg: 0.019833 BTC/block</span></p>
-<hr>
-<small>Generated • Fees from mempool.space • 2025-11-08 07:07:31</small>
-</div>
-</body>
-</html>
