--- output.html
@@ -1,22 +1,0 @@
-
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Bitcoin Mined Calculator</title>
-    <style>
-        body { font-family: Arial, sans-serif; padding: 40px; background: #fff8f0; }
-        h1 { color: #f7931a; }
-        .big { font-size: 36px; font-weight: bold; color: #00008b; }
-    </style>
-</head>
-<body>
-    <h1>Bitcoin Mined Calculator</h1>
-    <p>From block <strong>884,914</strong> to <strong>885,922</strong></p>
-    <p>Total blocks: <strong>1,009</strong></p>
-    <p class="big">3,153.12500000 BTC</p>
-    <p>Fees included: <strong>False</strong></p>
-    <p>Current block reward: <strong>3.12500000 BTC</strong></p>
-    <hr>
-    <small>Generated: 2025-11-08 05:01:51</small>
-</body>
-</html>
